initial move from Primer::HeadingComponent to Primer::Beta::Heading
