+++ app/components/primer/base_component.rb
@@ -14,7 +14,7 @@ module Primer
   # For example:
   #
   # ```erb
-  # <%= render Primer::HeadingComponent.new(mt: [0, nil, nil, 4, 2]) do %>
+  # <%= render Primer::Beta::Heading.new(mt: [0, nil, nil, 4, 2]) do %>
   #   Hello world
   # <% end %>
   # ```
