+++ static/audited_at.json
@@ -26,6 +26,7 @@
   "Primer::Beta::CloseButton": "",
   "Primer::Beta::Counter": "",
   "Primer::Beta::Flash": "",
+  "Primer::Beta::Heading": "",
   "Primer::Beta::Text": "",
   "Primer::Beta::Truncate": "",
   "Primer::Beta::Truncate::TruncateText": "",
