+++ static/statuses.json
@@ -26,6 +26,7 @@
   "Primer::Beta::CloseButton": "beta",
   "Primer::Beta::Counter": "beta",
   "Primer::Beta::Flash": "beta",
+  "Primer::Beta::Heading": "beta",
   "Primer::Beta::Text": "beta",
   "Primer::Beta::Truncate": "beta",
   "Primer::Beta::Truncate::TruncateText": "alpha",
@@ -47,7 +48,7 @@
   "Primer::DropdownMenuComponent": "deprecated",
   "Primer::FlexComponent": "deprecated",
   "Primer::FlexItemComponent": "deprecated",
-  "Primer::HeadingComponent": "beta",
+  "Primer::HeadingComponent": "deprecated",
   "Primer::HellipButton": "alpha",
   "Primer::HiddenTextExpander": "alpha",
   "Primer::IconButton": "beta",
