+++ stories/primer/beta/heading_stories.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+require "primer/beta/heading"
+
+class Primer::Beta::HeadingStories < ViewComponent::Storybook::Stories
+  layout "storybook_preview"
+
+  story(:heading) do
+    controls do
+      select(:tag, Primer::Beta::Heading::TAG_OPTIONS, Primer::Beta::Heading::TAG_FALLBACK)
+    end
+
+    content do
+      "This is a heading!"
+    end
+  end
+end
